+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # CHANGELOG
 
 ## 0.0.2dev
+* Renames `ploomber-engine` to `debug`
+* Adds `debuglater` engine
 
 ## 0.0.1 (2022-07-20)